change naming, add function
