+++ classifier/forms.py
@@ -1,0 +1,4 @@
+from django import forms
+
+class MyForm(forms.Form):
+    url = forms.CharField()
