+++ classifier/templates/top.html
@@ -1,0 +1,23 @@
+{% load static %}
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <title>Gunochan</title>
+    <link rel="stylesheet" href="{% static 'css/mystyle.css' %}">
+  </head>
+
+  <body>
+    {% block content %}
+    <div class="home-wrap">
+      <h1 class="home-title">Gunosy Classifier</h1>
+      <form action="result" method="POST">
+        {% csrf_token %}
+        <input id="id_url" class="home-input" maxlength="200" name="url" type="text" required="" placeholder="enter news article URL here">
+        <input class="home-go"type="submit" value="Go!"/>
+      </form>
+    </div>
+    {% endblock %}
+  </body>
+
+</html>
