--- django/myapp1/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class Myapp1Config(AppConfig):
-    name = 'myapp1'
