--- django/myapp1/forms.py
@@ -1,4 +1,0 @@
-from django import forms
-
-class MyForm(forms.Form):
-    url = forms.CharField()
