--- django/myapp1/static/css/mystyle.css
@@ -1,142 +1,0 @@
-html, body {
-  width: 100%; height: 100%;
-  margin: 0;
-}
-
-a {
-  text-decoration: none;
-  color: black;
-}
-
-
-.home-wrap {
-  width: 80%;
-  max-width: 400px;
-  margin: auto;
-}
-
-.home-title {
-  text-align: center;
-  padding-top: 100px;
-  font-size: 40px;
-}
-
-.home-input {
-  margin-top: 60px;
-  width: 100%;
-  font-size: 20px;
-  padding: 5px 10px;
-}
-.home-go {
-  display: block;
-  width: 80px;
-  margin: auto;
-  margin-top: 30px;
-  font-size: 16px;
-  text-align: center;
-  padding: 4px 10px;
-  border: 1px solid black;
-  background: #eee;
-  cursor: pointer;
-}
-
-header {
-  height: 40px;
-  width: 100%;
-  overflow: hidden;
-}
-.header-title {
-  float: left;
-  font-size: 24px;
-  font-weight: bold;
-  padding: 6px 14px;
-}
-
-.header-again{
-  float: right;
-  height: 24px;
-  padding: 8px;
-}
-
-.header-input {
-  height: 20px;
-  width: 400px;
-  padding-left: 4px;
-}
-.header-go {
-  height: 24px;
-  font-size: 16px;
-  padding: 2px 4px;
-  background: #eee;
-  margin-left: 0;
-}
-
-.result-wrap {
-  overflow: hidden;
-  width: 100%;
-  background: #eee;
-}
-
-.result-right {
-  float: right;
-  width: 40%;
-  margin: 10px 3.3% 20px 0;
-  padding: 10px 2.5% 20px;
-  background: white;
-}
-
-.result-left {
-  float: left;
-  width: 40%;
-  margin: 10px 3.3% 30px;
-  padding: 10px 2.5%;
-  background: white;
-}
-
-.category {
-  width: 90%;
-  margin: 10px 5%;
-}
-
-.category__head {
-  width: 100%;
-  overflow: hidden;
-  margin-bottom: 4px;
-}
-.category__head-name {
-  float: left;
-}
-
-.category__head-value {
-  float: right;
-}
-
-.category__body {
-  height: 4px;
-  width: 100%;
-  background: #eee;
-}
-.category__body-value {
-  height: 100%;
-  width: 40%;
-  background: blue;
-}
-
-@media screen and (max-width: 700px) {
- 
-.result-right {
-  float: none;
-  width: 85%;
-  margin: 10px 2.5% 20px;
-  padding: 10px 5% 20px;
-}
-
-.result-left {
-  float: none;
-  width: 90%;
-  margin: 10px 2.5% 30px;
-  padding: 10px 2.5%;
-}
-
-
-}
